backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
qdrant-client==1.8.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
cryptography>=41.0.0
markdown==3.5.1
beautifulsoup4==4.12.2